network.remove() + test refactoring
